# openapi/certifications-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Certifications.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/employees": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "List employees with server-side filtering and pagination",
        "operationId": "ListEmployees",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "name"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeSummaryDtoPagedResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Atomically create an employee and first contract",
        "operationId": "CreateEmployee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEmployeeResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/employees/{employeeId}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "summary": "Get employee details and the current contract",
        "operationId": "GetEmployee",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Employees"
        ],
        "summary": "Update editable employee data",
        "operationId": "UpdateEmployee",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/employees/{employeeId}/password/generate": {
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Generate and replace an employee password",
        "operationId": "GenerateEmployeePassword",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/employees/{employeeId}/password/reveal": {
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Reveal an employee password to an administrator",
        "operationId": "RevealEmployeePassword",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/me/password/reveal": {
      "post": {
        "tags": [
          "Passwords"
        ],
        "summary": "Reveal the current employee's password",
        "operationId": "RevealOwnPassword",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AdminMode": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "CertificationDto": {
        "type": "object",
        "properties": {
          "certificationId": {
            "type": "integer",
            "format": "int64"
          },
          "contractId": {
            "type": "integer",
            "format": "int64"
          },
          "assessor": {
            "type": "string",
            "nullable": true
          },
          "certificationDate": {
            "type": "string",
            "format": "date"
          },
          "protocolDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationSend": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationReturned": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "isCompleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CertificationStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ContractDetailsDto": {
        "type": "object",
        "properties": {
          "contract": {
            "$ref": "#/components/schemas/ContractDto"
          },
          "certifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CertificationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ContractDto": {
        "type": "object",
        "properties": {
          "contractId": {
            "type": "integer",
            "format": "int64"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "contractDate": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "effectiveValidTo": {
            "type": "string",
            "format": "date"
          },
          "active": {
            "type": "boolean"
          },
          "prolongationWarningMonths": {
            "type": "integer",
            "format": "int32"
          },
          "prolongationAlertMonths": {
            "type": "integer",
            "format": "int32"
          },
          "prolongationForYears": {
            "type": "integer",
            "format": "int32"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/CertificationStatus"
          }
        },
        "additionalProperties": false
      },
...
      "CreateContractRequest": {
        "type": "object",
        "properties": {
          "position": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "contractDate": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationWarningMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "prolongationAlertMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "prolongationForYears": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateEmployeeRequest": {
        "type": "object",
        "properties": {
          "personalId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "firstContract": {
            "$ref": "#/components/schemas/CreateContractRequest"
          }
        },
        "additionalProperties": false
      },
      "CreateEmployeeResultDto": {
        "type": "object",
        "properties": {
          "employee": {
            "$ref": "#/components/schemas/EmployeeDetailsDto"
          },
          "generatedPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EmployeeDetailsDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "personalId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "preferredAdminMode": {
            "$ref": "#/components/schemas/AdminMode"
          },
          "currentContract": {
            "$ref": "#/components/schemas/ContractDetailsDto"
          }
        },
        "additionalProperties": false
      },
      "EmployeeSummaryDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "personalId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "hasActiveContract": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/CertificationStatus"
          }
        },
        "additionalProperties": false
      },
      "EmployeeSummaryDtoPagedResult": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeSummaryDto"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "PasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UpdateEmployeeRequest": {
        "type": "object",
        "properties": {
          "personalId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Additional API-channel key required by every /api/v1 operation.",
        "name": "X-API-Key",
        "in": "header"
      },
      "CookieAuth": {
        "type": "apiKey",
        "description": "Secure HttpOnly authentication cookie issued by Login.",
        "name": "Certifications.Auth",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ],
...
}